Remove some duplicate code.

+++ cpp/dolfin/fem/utils.h
@@ -79,8 +79,8 @@ la::PETScVector
 create_vector_nest(const std::vector<const common::IndexMap*>& maps);
 
 /// Get new global index in 'spliced' indices
-std::size_t get_global_index(const std::vector<const common::IndexMap*>& maps,
-                             const int field, const int n);
+std::int64_t get_global_index(const std::vector<const common::IndexMap*>& maps,
+                              const int field, const int n);
 
 /// Create an ElementDofLayout from a ufc_dofmap
 ElementDofLayout create_element_dof_layout(const ufc_dofmap& dofmap,
